rename script for image pinning
