+++ workflows/helm_generator/image_pinning.py
